make keyboard-shortcuts-global a mixin
